backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
pillow==10.4.0
python-dotenv==1.0.1
timm==1.0.8
torch==2.10.0
torchvision==0.25.0
